requirements.txt
python-telegram-bot==12.2.0
Click==7.0
PyYAML==5.1.2
